mlapp/requirements.txt
flask
flask-cors
opencv-python-headless
mediapipe
numpy
pandas
scikit-learn
python-dotenv